Change the package structure
