+++ watson-personality-insights-cfenv-processor/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 io.pivotal.cfenv.spring.boot.CfEnvProcessor=\
-  org.terrence.cfenvprocessors.watsonpersonalityinsights.WatsonPersonalityInsightsCfEnvProcessor
+  com.ibm.boot.cfenvprocessors.watsonpersonalityinsights.WatsonPersonalityInsightsCfEnvProcessor
