+++ watson-tone-analyzer-cfenv-processor/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 io.pivotal.cfenv.spring.boot.CfEnvProcessor=\
-  org.terrence.cfenvprocessors.watsontoneanalyser.WatsonToneAnalyzerCfEnvProcessor
+  com.ibm.boot.cfenvprocessors.watsontoneanalyser.WatsonToneAnalyzerCfEnvProcessor
